+++ client/src/configuration/configuration.node.ts
@@ -1,0 +1,3 @@
+import { BaseServiceConfigurationProvider } from "./configuration";
+
+export class NodeServiceConfigurationProvider extends BaseServiceConfigurationProvider {}